+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### changed
 - Minimum requirement Doctrine ORM 2.13.
 - Doctrine XSD URLs for XML mapping.
+- `priority` is now a small int with a default value of 0.
 
 ## [4.0.0] - 2022-05-28
 ### changed